Box2ADLS/requirements.txt
# Freeze before deployment
adal
boxsdk
#jupyter
#numpy
openpyxl
#pandas
#pyodbc
pytest
requests

--pre
azure-storage-file-datalake
